Processes: Fix zygoteId() We cannot just extract SystemClock.elapsedRealtime() from System.currentTimeMillis() and return that as an ID. Both methods some times takes a few milliseconds to execute, which means that the first will be behind on the second and the calculation will change each time.
Instead we store the first calculation and check if the difference is above 3 seconds. This is because the methods will never take 3 seconds to execute, no mater how slow the device is, and the device will never boot in less than 3 seconds, no matter how fast it is. So if the difference is below 3 seconds, it is certain that the device has not been rebooted since the last calculation.